Tried to Rotate Chart
Yeahhh... that didn't work. If  you uncomment the first part of the
style tag, the chart DOES rotate, but it is drawn off the edge of the
browser and it's dumbfire: no rotation of text or anything else. The
other style attributes were an attempt at scaling but those only work on
SVG elements, not div elements.